refactor: all other files related to permission.config rename
