+++ pages/_prefix/explore/collectibles.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script lang="ts">
-import { explorerVisible } from '~/utils/config/permission.config'
+import { explorerVisible } from '@/utils/config/permission.config'
 
 export default {
   name: 'ExploreCollectibles',
